@@ -1,8 +1,12 @@
-import App from './app';
+import React from "react";
+import App from './components/App/App';
+import ReactDOM from 'react-dom/client';
+import './nullstyles.css';
 
-import ReactDOM from "react-dom";
-
-ReactDOM.render(<App />, document.getElementById("root"));
+const root = ReactDOM.createRoot(document.getElementById("root"))
+root.render(
+    <App />
+);
 
 const devMode = process.env.NODE_ENV === 'development';
 if (devMode && module && module.hot) {
